+++ examples/basic-auth/src/pages/About/index.tsx
@@ -1,0 +1,14 @@
+/* eslint jsx-a11y/aria-role: 0 */
+import React from 'react';
+import { Link } from 'ice';
+
+const About = () => {
+  return (
+    <div>
+      <h2>About Page</h2>
+      <Link to="/">Back Home</Link>
+    </div>
+  );
+};
+
+export default About;