+++ packages/build-app-helpers/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Changelog
 
-## 2.4.2
+## 2.5.2
 
 - [fix] support remote route
+
+## 2.5.1
+
+- [fix] slice error for array of source files
+
+## 2.5.0
+
+- [chore] revert `@babel/plugin-transform-runtime` helpers value
+- [chore] bump version of `es-module-lexer`
+- [feat] support scan source code of import statement
+
+## 2.4.3
+
+- [chore] not inline @babel/runtime/helpers code
+
+## 2.4.2
+
+- [fix] import redirection with alias
 
 ## 2.4.1