+++ packages/build-app-templates/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 1.1.3
+
+- [fix] relative path of `loadStaticModules`
+
+## 1.1.2
+
+- [fix] rax routerAPI template
+
+## 1.1.1
+
+- [fix] `appConfig` need a default value
+- [fix] rax runApp template export `getSearchParams` method
+
 ## 1.1.0
 
 - [chore] use `enableRouter` replace `buildConfig.router`