+++ packages/build-mpa-config/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 4.2.0
+
+- [feat] support disable runtime before render template
+- [fix] file path render in ejs template
+
+## 4.1.3
+
+- [fix] fix: format mpa `app.json` path in windows
+
+## 4.1.2
+
+- [fix] format path in case of win32 system
+- [fix] assign `routesFilePath` in `RaxGenerator`
+
 ## 4.1.1
 
 - [feat] support frm mode