+++ packages/build-user-config/CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## 2.1.0
+
+- [feat] support `--force` to empty cache folder
+- [fix] swc use esm type
+
+## 2.0.5
+
+- [fix] `process.env.__IS_SERVER__` default value is `false`
+
+## 2.0.4
+
+- [fix] `minify: true` is invalid
+
+## 2.0.3
+
+- [fix] lazy import is invalid with swc
+
+## 2.0.2
+
+- [fix] compatible with alias which config with exists relative file path
+
+## 2.0.1
+
+- [fix] enhance alias for node_modules dependencies
+
+## 2.0.0
+
+- [fix] add url fallback for webpackHotDevClient
+- [fix] bump version of `@builder/pack`(^0.5.0)
+- [feat] add https url list
+- [chore] enhance `applyCliOptions`
+- [chore] RedirectPathLoader should include `.rax|.ice`
+- [fix] skip deps `filepath` in windows
+
 ## 1.1.9
 
 - [feat] compact compact postcss options in different version of postcss-loader