+++ packages/plugin-ice-ssr/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.1.0
+
+- [feat] support SSR in mode Vite
+- [feat] support i18n generation
+
+## 3.0.6
+
+- [fix] server-side render failed in production(always render `global.__ICE_SERVER_HTML_TEMPLATE__`)
+- [chore] remove `build/server/loadable-stats.json`
+
 ## 3.0.5
 
 - [fix] update dependencies version