+++ packages/plugin-mpa/CHANGELOG.md
@@ -1,5 +1,13 @@
 # changelog
 
+## 2.1.0
+
+- [feat] optimize runtime when build
+
+## 2.0.4
+
+- [fix] format path in case of win32 system
+
 ## 2.0.3
 
 - [fix] update dependencies version