+++ packages/plugin-router/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.1
+
+- [fix] bump version of `@builder/pack`(^0.5.0)
+
 ## 2.1.0
 
 - [chore] migrate wrapper ErrorBoundary/SearchParams/SSR/CSR to plugin-react-app