+++ packages/plugin-stark-module/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## 2.2.4
+
+- [fix] when `url` configured, compile externals to `runtime.json`.
+- [fix] add dev log.
+
 ## 2.2.3
 
-- [fix] move `webpack-plugin-import` to dependencies
+- [fix] move `webpack-plugin-import` to dependencies.
 
 ## 2.2.2