+++ packages/vite-plugin-component-style/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.4
+
+- [chore] bump version of `es-module-lexer`
+
+## 1.0.3
+
+- [fix] style import for multiple components with `stylePath`
+
 ## 1.0.2
 
 - [fix] deprecate check of `pkg.componentConfig`